ResourceHelper getOrCreate, python wrappers
